Efficiency updates and concurrency correction
+ Changed render thread setup to select a particular loop based on settings rather than testing every frame
+ changed vertexShaders arrayList in RenderThread class to CopyOnWriteArrayList to resolve some concurrency issues. May change if efficiency becomes an issue

+++ slythr/Line.java
@@ -126,7 +126,7 @@ public class Line extends Primitive {
      * @param time Scale the distance by this factor
      */
     public void move(double time){
-        setpos((double) getpos()[0] + (physics_velocity_x * time), (double) getpos()[1] + (physics_velocity_y * time));
+        setpos(getpos()[0] + (physics_velocity_x * time), getpos()[1] + (physics_velocity_y * time));
     }
 
     /**
